fixed PropertyService logic for property report

+++ src/main/java/com/propify/challenge/model/PropertyReport.java
@@ -9,9 +9,9 @@ public class PropertyReport {
 
     Integer totalQuantity;
 
-    Map<PropertyType, Integer> quantityPerType;
+    Map<PropertyType, Long> quantityPerType;
 
     double averageRentPrice;
 
-    Integer illinoisQuantity;
+    Long illinoisQuantity;
 }
